added locks and changed userid datatype

+++ repository/repository.go
@@ -8,7 +8,7 @@ import (
 
 type Repo interface {
 	Add(ctx context.Context, user models.User) (err error)
-	Delete(ctx context.Context, userId int32) (deletedUser models.User, err error)
+	Delete(ctx context.Context, userId int64) (deletedUser models.User, err error)
 	Update(ctx context.Context, user models.User) (updatedUser models.User, err error)
 	GetUsers(ctx context.Context) (users []models.User, err error)
 }
